fix windows support for other palm-generate switches

+++ vscode-generate.bat
@@ -1,21 +1,23 @@
 SETLOCAL EnableDelayedExpansion
 set args="%*"
-if x%args:new_scene=%==x%args% (
-    call :lastarg %*
-    set "PROJECT_PATH=!LAST_ARG!"
-    if exist "!PROJECT_PATH!" (
-        echo adding visualstudio code extensions
-        set "VSCODE_DIR=!PROJECT_PATH!\.vscode"
-        set "BIN_DIR=!PROJECT_PATH!\_bin"
-        set "EXT_DIR=%HERE%..\vscode"
-        mkdir "!VSCODE_DIR!"
-        mkdir "!BIN_DIR!"
-        copy "!EXT_DIR!\_scripts\*" "!VSCODE_DIR!\" >nul
-        copy "!EXT_DIR!\tasks.json" "!VSCODE_DIR!\" >nul
-        REM Tell Windows not to block script execution
-        echo.>"!VSCODE_DIR!\build.bat":Zone.Identifier
-        echo.>"!VSCODE_DIR!\run.bat":Zone.Identifier
-    )
+if "%~1" NEQ "" (
+	if x%args:new_scene=%==x%args% (
+		call :lastarg %*
+		set "PROJECT_PATH=!LAST_ARG!"
+		if exist "!PROJECT_PATH!" (
+			echo adding visualstudio code extensions
+			set "VSCODE_DIR=!PROJECT_PATH!\.vscode"
+			set "BIN_DIR=!PROJECT_PATH!\_bin"
+			set "EXT_DIR=%HERE%..\vscode"
+			mkdir "!VSCODE_DIR!"
+			mkdir "!BIN_DIR!"
+			copy "!EXT_DIR!\_scripts\*" "!VSCODE_DIR!\" >nul
+			copy "!EXT_DIR!\tasks.json" "!VSCODE_DIR!\" >nul
+			REM Tell Windows not to block script execution
+			echo.>"!VSCODE_DIR!\build.bat":Zone.Identifier
+			echo.>"!VSCODE_DIR!\run.bat":Zone.Identifier
+		)
+	)
 )
 
 :lastarg
